backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic[email]==2.6.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
